Added bounding box selection

+++ Controller/Controller.h
@@ -25,7 +25,7 @@ public:
     void generateCylinder(const string&, bool, float, float, float);
     void generateCone(const string&, bool, float, float, float);
 
-    void translate(int,int,int);
+    void translate(float,float,float);
     void rotate(float,char);
     void tmp();
 };
